+++ Patches/ObjectDBPatch.cs
@@ -1,30 +1,29 @@
-﻿//using HarmonyLib;
-//using UnityEngine.SceneManagement;
+﻿using HarmonyLib;
+using UnityEngine.SceneManagement;
 
 
-//namespace MoreVanillaBuildPrefabs
-//{
-//    // Switch to using Azumatt piece manager instead of Jotunn?
+namespace MoreVanillaBuildPrefabs
+{
 
-//    [HarmonyPatch(typeof(ObjectDB))]
-//    internal class ObjectDBPatch
-//    {
-//        // Hook just before Jotunn registers the Pieces
-//        [HarmonyPrefix]
-//        [HarmonyPatch(nameof(ObjectDB.Awake))]
-//        static void ObjectDBAwake()
-//        {
-//#if DEBUG
-//            Log.LogInfo("ObjectDBAwake()");
-//#endif
-//            if (PluginConfig.IsModEnabled.Value)
-//            {
-//                if (SceneManager.GetActiveScene().name == "start" && PrefabHelper.AddedPieces.Count != 0)
-//                {
-//                    PrefabHelper.RemoveAddedPrefabs();
-//                    Plugin.RemoveHammerCategories();
-//                }
-//            }
-//        }
-//    }
-//}
+    [HarmonyPatch(typeof(ObjectDB))]
+    internal class ObjectDBPatch
+    {
+        // Hook just before Jotunn registers the Pieces
+        [HarmonyPrefix]
+        [HarmonyPatch(nameof(ObjectDB.Awake))]
+        static void ObjectDBAwake()
+        {
+#if DEBUG
+            Log.LogInfo("ObjectDBAwake()");
+#endif
+            if (PluginConfig.IsModEnabled.Value)
+            {
+                if (SceneManager.GetActiveScene().name == "start" && PrefabHelper.AddedPrefabs.Count != 0)
+                {
+                    PrefabHelper.RemoveCustomPieces();
+                    Hammer.RemoveHammerCategories();
+                }
+            }
+        }
+    }
+}